Contribution: User can view file changes side-by-side
Users can view changes made to files within a contribution side-by-side.

Resolves [#275](https://github.com/OpenlyOne/openly/issues/275).
