+++ app/controllers/abstract/file_infos_controller.rb
@@ -22,7 +22,7 @@ module Abstract
     def index
       render 'file_infos/index', locals: {
         path_parameters: path_parameters,
-        file_infos_path: "#{link_path_prefix}file_infos_path",
+        file_change_path: "#{link_path_prefix}file_change_path",
         folder_path: "#{link_path_prefix}folder_path",
         force_syncs_path: "#{link_path_prefix}force_syncs_path",
         root_folder_path: "#{link_path_prefix}root_folder_path"
