+++ app/views/file_infos/index.slim
@@ -8,6 +8,7 @@
                        user_can_force_sync_files: @user_can_force_sync_files,
                        project: @project,
                        path_parameters: path_parameters,
+                       file_change_path: file_change_path,
                        folder_path: folder_path,
                        force_syncs_path: force_syncs_path,
                        root_folder_path: root_folder_path }
