+++ app/views/revisions/_file_changes.slim
@@ -10,6 +10,8 @@
              collection: file_changes,
              locals: { project: project,
                        revision: revision,
+                       path_parameters: path_parameters,
+                       file_change_path: file_change_path,
                        show_link_to_file_info: show_link_to_file_info,
                        all_links_in_new_tab: all_links_in_new_tab }
   - else
