+++ app/views/revisions/index.slim
@@ -4,6 +4,7 @@
 
   = render partial: 'revision',
            collection: @revisions,
-           locals: { project: @project, show_link_to_file_info: true }
+           locals: { project: @project,
+                     show_link_to_file_info: true }
 
 .spacing.v48px
