+++ config/routes.rb
@@ -108,6 +108,9 @@ Rails.application.routes.draw do
           # Routes for file infos
           resources :file_infos, path: 'files/:id/info', only: :index
           resources :force_syncs, path: 'files/:id/sync', only: :create
+
+          # Side-by-side diffs
+          resources :file_changes, path: 'changes', only: :show
         end
       end
